Moved queue related code closer to where it is used or where its use matches the file name better.
