+++ platform/queue/builtin/db.go
@@ -10,17 +10,21 @@ import (
 	"github.com/micromdm/micromdm/mdm"
 )
 
+const (
+	DeviceCommandBucket = "mdm.DeviceCommands"
+)
+
 type DB struct {
 	*bolt.DB
 }
 
 func NewDB(db *bolt.DB) (*DB, error) {
 	err := db.Update(func(tx *bolt.Tx) error {
-		_, err := tx.CreateBucketIfNotExists([]byte(queue.DeviceCommandBucket))
+		_, err := tx.CreateBucketIfNotExists([]byte(DeviceCommandBucket))
 		return err
 	})
 	if err != nil {
-		return nil, errors.Wrapf(err, "creating %s bucket", queue.DeviceCommandBucket)
+		return nil, errors.Wrapf(err, "creating %s bucket", DeviceCommandBucket)
 	}
 	return &DB{DB: db}, nil
 }
@@ -30,9 +34,9 @@ func (db *DB) Save(ctx context.Context, cmd *queue.DeviceCommand) error {
 	if err != nil {
 		return errors.Wrap(err, "begin transaction")
 	}
-	bkt := tx.Bucket([]byte(queue.DeviceCommandBucket))
+	bkt := tx.Bucket([]byte(DeviceCommandBucket))
 	if bkt == nil {
-		return fmt.Errorf("bucket %q not found!", queue.DeviceCommandBucket)
+		return fmt.Errorf("bucket %q not found!", DeviceCommandBucket)
 	}
 	devproto, err := queue.MarshalDeviceCommand(cmd)
 	if err != nil {
@@ -48,7 +52,7 @@ func (db *DB) Save(ctx context.Context, cmd *queue.DeviceCommand) error {
 func (db *DB) DeviceCommand(ctx context.Context, udid string) (*queue.DeviceCommand, error) {
 	var dev queue.DeviceCommand
 	err := db.View(func(tx *bolt.Tx) error {
-		b := tx.Bucket([]byte(queue.DeviceCommandBucket))
+		b := tx.Bucket([]byte(DeviceCommandBucket))
 		v := b.Get([]byte(udid))
 		if v == nil {
 			return &queue.NotFound{ResourceType: "DeviceCommand", Message: fmt.Sprintf("udid %s", udid)}
