+++ platform/queue/device_command.go
@@ -1,6 +1,7 @@
 package queue
 
 import (
+	"fmt"
 	"time"
 
 	"github.com/gogo/protobuf/proto"
@@ -171,4 +172,20 @@ func UnmarshalDeviceCommand(data []byte, c *DeviceCommand) error {
 		})
 	}
 	return nil
+}
+
+type NotFound struct {
+	ResourceType string
+	Message      string
+}
+
+func (e *NotFound) Error() string {
+	return fmt.Sprintf("not found: %s %s", e.ResourceType, e.Message)
+}
+
+func IsNotFound(err error) bool {
+	if _, ok := err.(*NotFound); ok {
+		return true
+	}
+	return false
 }
